+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "test": "npm run start:prod && jest",
+    "test": "jest",
     "start:dev": "nodemon --experimental-specifier-resolution=node ./src/index.ts",
     "start:prod": "webpack && node ./dist/bundle.cjs"
   },
