+++ tests/server.test.ts
@@ -58,35 +58,35 @@ describe('Server', () => {
   })
 });
 
-describe('Server handles invalid request:', () => {
-  beforeAll(async()=>{
-    const res = await fetch(`http://localhost:${PORT}/api/users`, {
-      method: "POST",
-      body: JSON.stringify(testUser_1)
-    });
-    const data = await res.json();
-    userId = data.id;
-  });
-  it.each(INVALID_ENDPOINTS)('with non-existing endpoint %s returns 404', async (endpoint) => {
-    const res = await fetch(`http://localhost:${PORT}${endpoint}`);
-    expect(res.status).toBe(404);
-  });
-  it.each(INVALID_UPDATES)('with invalid user %o returns error msg', async (payload) => {
-    const res = await fetch(`http://localhost:${PORT}/api/users/${userId}`, {
-      method: 'PUT',
-      body: JSON.stringify(payload)
-    });
-    const msg = await res.text();
-    expect(msg).toBe('Check provided info');
-  });
-  it('with invalid uuid', async () => {
-    const fakeUUID = v4();
-    const res = await fetch(`http://localhost:${PORT}/api/users/someFakeUUID`);
-    expect(res.status).toEqual(400);
-  })
-  it('with non-existing userId', async () => {
-    const fakeUUID = v4();
-    const res = await fetch(`http://localhost:${PORT}/api/users/${fakeUUID}`);
-    expect(res.status).toBe(404);
-  }) 
-});
+// describe('Server handles invalid request:', () => {
+//   beforeAll(async()=>{
+//     const res = await fetch(`http://localhost:${PORT}/api/users`, {
+//       method: "POST",
+//       body: JSON.stringify(testUser_1)
+//     });
+//     const data = await res.json();
+//     userId = data.id;
+//   });
+//   it.each(INVALID_ENDPOINTS)('with non-existing endpoint %s returns 404', async (endpoint) => {
+//     const res = await fetch(`http://localhost:${PORT}${endpoint}`);
+//     expect(res.status).toBe(404);
+//   });
+//   it.each(INVALID_UPDATES)('with invalid user %o returns error msg', async (payload) => {
+//     const res = await fetch(`http://localhost:${PORT}/api/users/${userId}`, {
+//       method: 'PUT',
+//       body: JSON.stringify(payload)
+//     });
+//     const msg = await res.text();
+//     expect(msg).toBe('Check provided info');
+//   });
+//   it('with invalid uuid', async () => {
+//     const fakeUUID = v4();
+//     const res = await fetch(`http://localhost:${PORT}/api/users/someFakeUUID`);
+//     expect(res.status).toEqual(400);
+//   })
+//   it('with non-existing userId', async () => {
+//     const fakeUUID = v4();
+//     const res = await fetch(`http://localhost:${PORT}/api/users/${fakeUUID}`);
+//     expect(res.status).toBe(404);
+//   }) 
+// });
